test(config-resolver): fold three single-case test files into parent suite (refs #4122)
Fold tests/lib/config-resolver-commands.test.js,
config-resolver-worktree.test.js, and config-resolver-notifications.test.js
into tests/lib/config-resolver.test.js. The three absorbed cases (commands
defaults, worktreeIsolation defaults, notifications defaults) move verbatim
— every assertion and explanatory comment preserved — and now execute as
the final three suites of the parent file (29 → 32 tests). Extends the
parent import list with NOTIFICATIONS_DEFAULTS, resolveWorktreeEnabled, and
WORKTREE_ISOLATION_DEFAULTS, then deletes the three now-empty files. Net:
zero coverage loss, three fewer files — trimming the over-fragmented corner
the review flagged.
