--- tests/lib/config-resolver-commands.test.js
@@ -1,45 +1,0 @@
-import assert from 'node:assert/strict';
-import path from 'node:path';
-import { beforeEach, describe, it } from 'node:test';
-import { Volume } from 'memfs';
-import {
-  COMMANDS_DEFAULTS,
-  PROJECT_ROOT,
-  resolveConfig,
-} from '../../.agents/scripts/lib/config-resolver.js';
-import { setupFsMock } from './fs-mock.js';
-
-/**
- * Regression: callers that read `config.project.commands.test` directly
- * (without `getCommands()`) previously got `undefined` when the operator
- * omitted the `commands` block. `applyDefaults()` now enriches the
- * canonical block with COMMANDS_DEFAULTS so every field is present.
- */
-describe('config-resolver — commands defaults', () => {
-  let vol;
-
-  beforeEach((t) => {
-    vol = new Volume();
-    setupFsMock(t, vol);
-    resolveConfig({ bustCache: true });
-  });
-
-  it('applies COMMANDS_DEFAULTS when project.commands is omitted', () => {
-    vol.mkdirSync(PROJECT_ROOT, { recursive: true });
-    vol.writeFileSync(
-      path.join(PROJECT_ROOT, '.agentrc.json'),
-      JSON.stringify({
-        project: {
-          paths: { agentRoot: '.agents', docsRoot: 'docs', tempRoot: 'temp' },
-        },
-      }),
-    );
-    // Epic #2880 / F14B removed `agentSettings` from the resolved config
-    // (hard-cutover of the legacy shim per `git-conventions.md`). The
-    // canonical surface for these defaults is `project.commands`.
-    const { project } = resolveConfig({ bustCache: true });
-    assert.equal(project.commands.test, COMMANDS_DEFAULTS.test);
-    assert.equal(project.commands.lintBaseline, COMMANDS_DEFAULTS.lintBaseline);
-    assert.equal(project.commands.formatCheck, COMMANDS_DEFAULTS.formatCheck);
-  });
-});
